requirements.txt
Flask==2.2.3
pdfminer.six==20221105
openai==0.27.0
Werkzeug==2.2.3
python-dotenv==1.0.0
